v0.0.5
Okay lets see, where to start...
( + ) = New Content,  ( o ) = Modified Content,  ( - ) = Removed Content

( + ) Users automatically connect to the server now when the application is launched, but they are not put into any lobbies or rooms.
( + ) Login Panel created in Lobby Scene.  User can designate their own name now, and log in to the main lobby using that name.  Users must designate a name and cannot leave this blank.  Currently there is no type-case restrictions, but will be implemented at a later date.
( o ) Modified the appearance of the Lobby Panel, which now also includes a placeholder chat box.  I tried implementing a chat in the Lobby, but due to certain restrictions, I was unable to do so.
( o ) Slightly modified the appearance of the Room Creation Panel.  No other changes made here.
( o ) Reworked the appearance of the current Room Info Panel.  It now uses a horizontal display rather than a vertical display.  The start button is now made uninteractable by regular clients and is replaced with "Waiting For Host" text.
( + ) Room Info Panel now has a functional chat implemented into it.  While I couldn't make it work for the lobby due to restrictions on how the server works, I managed to get it working when the player is in a room.
( o ) The master client of the room can now start the match for each player in the room by hitting the Join Game button.  This results in a lobby-to-game scene transition, in which each player loads into the devTower scene.
( + ) Added a Logout button so players can leave the lobby and return to the Login Panel
( + ) Added a Quit button that closes the application from anywhere.
( + ) Added two text displays in the bottom right of the screen that display your current connection status and the current version/build of the application.
( - ) Removed the NetworkPlayer script.  As I move into learning how to use server properties to hold data, I no longer need to store it in a separate class upon connecting to the server.
( + ) Created a Networked Player Controller in the game scene.  Only the client that instantiated it is able to access its controls, but any transformations made in game are reflected on other clients screens (albeit with significant lag).
( + ) Created Networked Game Manager in the game scene.  This class is responsible for instantiating the player when they load in.  It also tracks the victory condition and ends the match when it has been met.
( + ) Added victory condition:  first player to touch the food at the top of the tower wins the game.
( o ) Expanded the tower map significantly, introducing 7 "biomes" which will be used for testing mechanics.  This biomes are as listed:  Grasslands, Sand/Desert, Forest/Jungle, Ice/FrozenPeaks, Fire/Lava, Water/Ocean, Cityscape.
( ? ) I believe the bug where multiple rooms being displayed got fixed.  Not 100% on this yet, but I didn't see it occur during my testing.
( + ) Implemented a single player debug mode for the game scene, that lets you test things without connecting to the server.
( o ) Modified the CameraFollow script so that it tracks only the client player object, rather than the first player object that spawned in.
( + ) Implemented a victoy screen when victory condition is met in the game scene.  However there is currently no way to back out of this, meaning you cannot return back to lobby scene, you cannot restart the match, and you cannot hide the victory canvas.  This will be modified in the future when game modes and the game-to-lobby scene transition is implemented.
( - ) Removed the old player prefabs in favor of the new SPController prefab.