v0.0.5 - Hotfix
( - ) Cleared the test chat from the room info chat box.

Notable Bug:  Whenever you leave a room, the OnConnectedToMaster() server function is called, resulting in the Login page being opened.  Consider moving the Login Panel display to a separate function or wrapping the code in a ( IsConnectedToMaster ) conditional statement to see if the player is being kicked out of the server and re-entered with the same information.